Move list of documentTypes and related types into index file
